NET-276: Use long instead of int for newsgroup article counts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/branches/NET_2_0@771683 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/examples/nntp/ExtendedNNTPOps.java
@@ -58,8 +58,8 @@ public class ExtendedNNTPOps {
             // XOVER
             NewsgroupInfo testGroup = new NewsgroupInfo();
             client.selectNewsgroup("alt.test", testGroup);
-            int lowArticleNumber = testGroup.getFirstArticle();
-            int highArticleNumber = lowArticleNumber + 100;
+            long lowArticleNumber = testGroup.getFirstArticle();
+            long  highArticleNumber = lowArticleNumber + 100;
             List<Article> articles = NNTPUtils.getArticleInfo(client, lowArticleNumber, highArticleNumber);
 
             for (Article article : articles) {
